requirements.txt
streamlit>=1.34.0
ultralytics>=8.1.0
torch>=2.0.0
torchvision>=0.15.0
pandas>=2.0.0
Pillow>=9.0.0
altair>=5.0.0
numpy>=1.24.0
